Initial Trekker Server add + WIP on Trekker Flatfish
Code is still very much a WIP here.  I stripped out a lot of the Trekker Flatfish guts
b/c we just don't need it for this static use case.

This code won't run, but the end result should be a lot simpler and more
straightforward should we need to extend or modify it in the future.

TODO work on config+ctools exportable, .install, and mappings.